refactor(webhook): extract signature validation into webhook_auth mixin
